+++ src/ska_low_mccs_spshw/tile/tile_component_manager.py
@@ -1172,7 +1172,7 @@ class TileComponentManager(
     def _subrack_says_state_changed(
         self: TileComponentManager,
         event_name: str,
-        event_value: list[float] | None,
+        event_value: tango.DevState | None,
         event_quality: tango.AttrQuality,
     ) -> None:
         """